+++ src/luascript.h
@@ -400,6 +400,7 @@ protected:
 	void registerGameFunctions();
 	void registerPositionFunctions();
 	void registerTileFunctions();
+	void registerNetworkMessageFunctions();
 	void registerPlayerFunctions();
 
 	lua_State* luaState = nullptr;
@@ -507,34 +508,6 @@ private:
 	static int luaVariantGetNumber(lua_State* L);
 	static int luaVariantGetString(lua_State* L);
 	static int luaVariantGetPosition(lua_State* L);
-
-	// NetworkMessage
-	static int luaNetworkMessageCreate(lua_State* L);
-	static int luaNetworkMessageDelete(lua_State* L);
-
-	static int luaNetworkMessageGetByte(lua_State* L);
-	static int luaNetworkMessageGetU16(lua_State* L);
-	static int luaNetworkMessageGetU32(lua_State* L);
-	static int luaNetworkMessageGetU64(lua_State* L);
-	static int luaNetworkMessageGetString(lua_State* L);
-	static int luaNetworkMessageGetPosition(lua_State* L);
-
-	static int luaNetworkMessageAddByte(lua_State* L);
-	static int luaNetworkMessageAddU16(lua_State* L);
-	static int luaNetworkMessageAddU32(lua_State* L);
-	static int luaNetworkMessageAddU64(lua_State* L);
-	static int luaNetworkMessageAddString(lua_State* L);
-	static int luaNetworkMessageAddPosition(lua_State* L);
-	static int luaNetworkMessageAddDouble(lua_State* L);
-	static int luaNetworkMessageAddItem(lua_State* L);
-	static int luaNetworkMessageAddItemId(lua_State* L);
-
-	static int luaNetworkMessageReset(lua_State* L);
-	static int luaNetworkMessageSeek(lua_State* L);
-	static int luaNetworkMessageTell(lua_State* L);
-	static int luaNetworkMessageLength(lua_State* L);
-	static int luaNetworkMessageSkipBytes(lua_State* L);
-	static int luaNetworkMessageSendToPlayer(lua_State* L);
 
 	// ModalWindow
 	static int luaModalWindowCreate(lua_State* L);
